+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The [Unreleased](#unreleased) section collects notes for unreleased changes and 
 Note that not every tagged version may be suitable for production use. A Github
 release will be created for any release tested in production, and may be marked below with the tag [Release]
 
+## Unreleased
+
+- [Added] :current_user_roles SQL substitution to provide an array[] of current user active role names
+- [Added] documentation for report SQL substitutions
+- [Fixed] filtering of config libraries to include name, so we can directly link to them from dynamic def admin pages
+
 ## [8.2.99] - 2023-10-18
 
 - [Added] the ability for reference action captions to incorporate more extensive substitutions